textarea fix

+++ reviews/forms.py
@@ -6,7 +6,7 @@ class ReviewForm(forms.ModelForm):
     name = forms.CharField(max_length=100)
     email = forms.EmailField()
     review_text = forms.CharField(widget=forms.Textarea)
-    rating = forms.IntegerField(min_value=1, max_value=10)
+    rating = forms.FloatField(min_value=1, max_value=10)
     captcha = ReCaptchaField()
 
     class Meta:
